requirements.txt
arxiv
langgraph
langchain
langchain-community
sentence-transformers
scikit-learn
jinja2
python-dotenv
pydantic
